new ITargetable, turrets now use triggerStay, new buff vfx manager, fixed bug in buff manager about new stats, projectile movement implemented, tracking movement implemented,  projectile redesigned to use hurtbox

+++ Assets/_Scripts/Helpers/NumberShortHand.cs
@@ -31,7 +31,7 @@ public static class NumberShorthand
         }
         else
         {
-            return num.ToString();
+            return ((int)num).ToString();
         }
     }
 }
